src/requirements.txt
Requests==2.32.3

cutie~=0.3.2
colorama~=0.4.6
pyfiglet~=1.0.2
PyOpenGL~=3.1.7
numpy~=2.1.2
pillow~=11.0.0
Cython~=3.0.11
jsonpickle~=3.3.0
pathlib~=1.0.1
Pygments~=2.18.0
python-dotenv~=1.0.1
tabulate~=0.9.0